fix Selection.getRng for firefox under special circumstances
In special cases, the 'lastFocusBookmark' in Firefox does not contain the correct element (though 'selectedRange' still points to the range with the correct element). This causes Exceptions in Selection.getRng.

The code in Selection.getRng has a IE bugfix conditional comment, but did not restrict it to only be used in IE. Doing so fixes all problems for Firefox.